Split IntegrationsSection into sub-modules (1,207 → 253 lines)
Each integration kind now lives in its own file under
components/admin/sections/integrations/:

- shared.tsx               types, constants, FieldRow, EmptyIntegrationCard,
                           account label helpers, save-button label helper
- AiHostPanel.tsx          AI host status + test panel + test state type
- SimpleIntegrationPanel   generic key/value panel for non-account
                           integrations (e.g. firewall)
- MultiAccountPanel        accounts list shell + collapsible AccountRow
- account-forms.tsx        per-type forms (GitHub / Email / Google)
- CustomIntegrationsPanel  list shell + add/save controls
- CustomIntegrationRow     collapsible row + field editor

IntegrationsSection.tsx is now the data-owning shell that holds draft
state, handles fetch/save/test, and dispatches to the right panel.
All sub-files are under 260 lines.
